queryDSL update

+++ src/main/java/com/shhwang0930/mytg/board/controller/BoardController.java
@@ -1,7 +1,6 @@
 package com.shhwang0930.mytg.board.controller;
 
 import com.shhwang0930.mytg.board.model.BoardDTO;
-import com.shhwang0930.mytg.board.model.BoardEntity;
 import com.shhwang0930.mytg.board.service.BoardService;
 import com.shhwang0930.mytg.common.model.ResponseMessage;
 import com.shhwang0930.mytg.common.model.StatusCode;
@@ -9,6 +8,8 @@ import com.shhwang0930.mytg.common.model.StatusCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -50,11 +51,21 @@ public class BoardController {
 
     @PutMapping("/me/{idx}")
     public ResponseEntity<ResponseMessage> updateBoard(@PathVariable Long idx, @RequestBody BoardDTO boardDTO){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+
+        if(!boardService.matchUser(username, idx)){
+            StatusCode statusCode = StatusCode.USER_MATCH_FAILED;
+            ResponseMessage responseMessage = new ResponseMessage(statusCode.getCode(), statusCode.getMessage(), null);
+            return ResponseEntity.status(HttpStatus.OK).body(responseMessage);
+        }
+
         if(!boardService.boardIsExist(idx)){
             StatusCode statusCode = StatusCode.BOARD_NOT_FOUND;
             ResponseMessage responseMessage = new ResponseMessage(statusCode.getCode(), statusCode.getMessage(), null);
             return ResponseEntity.status(HttpStatus.OK).body(responseMessage);
         }
+
         StatusCode statusCode = StatusCode.SUCCESS;
         ResponseMessage responseMessage = new ResponseMessage(statusCode.getCode(), statusCode.getMessage(), null);
         boardService.updateBoard(idx, boardDTO);
@@ -63,11 +74,22 @@ public class BoardController {
 
     @DeleteMapping("/me/{idx}")
     public ResponseEntity<ResponseMessage> deleteBoard(@PathVariable Long idx){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+
+        if(!boardService.matchUser(username, idx)){
+            StatusCode statusCode = StatusCode.USER_MATCH_FAILED;
+            ResponseMessage responseMessage = new ResponseMessage(statusCode.getCode(), statusCode.getMessage(), null);
+            return ResponseEntity.status(HttpStatus.OK).body(responseMessage);
+        }
+
+
         if(!boardService.boardIsExist(idx)){
             StatusCode statusCode = StatusCode.BOARD_NOT_FOUND;
             ResponseMessage responseMessage = new ResponseMessage(statusCode.getCode(), statusCode.getMessage(), null);
             return ResponseEntity.status(HttpStatus.OK).body(responseMessage);
         }
+
         StatusCode statusCode = StatusCode.SUCCESS;
         ResponseMessage responseMessage = new ResponseMessage(statusCode.getCode(), statusCode.getMessage(), null);
         boardService.deleteBoard(idx);
